+++ cursive-core/src/views/checkbox.rs
@@ -12,7 +12,7 @@ use std::rc::Rc;
 /// # Examples
 ///
 /// ```
-/// use cursive_core::traits::Identifiable;
+/// use cursive_core::traits::Nameable;
 /// use cursive_core::views::Checkbox;
 ///
 /// let checkbox = Checkbox::new().checked().with_name("check");