+++ cursive-core/src/views/linear_layout.rs
@@ -17,7 +17,7 @@ use std::ops::Deref;
 /// # Examples
 ///
 /// ```
-/// use cursive_core::traits::Boxable;
+/// use cursive_core::traits::Resizable;
 /// use cursive_core::views::{Button, LinearLayout, TextArea, TextView};
 ///
 /// let linear_layout = LinearLayout::horizontal()