+++ cursive-core/src/views/mod.rs
@@ -136,19 +136,3 @@ pub use self::{
     themed_view::ThemedView,
     tracked_view::TrackedView,
 };
-
-/// Same as [`LastSizeView`](self::LastSizeView).
-#[deprecated(note = "`SizedView` is being renamed to `LastSizeView`")]
-pub type SizedView<T> = LastSizeView<T>;
-
-/// Same as [`ResizedView`](self::ResizedView).
-#[deprecated(note = "`BoxView` is being renamed to `ResizedView`")]
-pub type BoxView<T> = ResizedView<T>;
-
-/// Same as [`BoxedView`](self::BoxedView).
-#[deprecated(note = "`ViewBox` is being renamed to `BoxedView`")]
-pub type ViewBox = BoxedView;
-
-/// Same as [`NamedView`](self::NamedView).
-#[deprecated(note = "`IdView` is being renamed to `NamedView`")]
-pub type IdView<T> = NamedView<T>;