+++ cursive/examples/colors.rs
@@ -1,5 +1,5 @@
 use cursive::theme::{Color, ColorStyle};
-use cursive::view::Boxable;
+use cursive::view::Resizable;
 use cursive::views::Canvas;
 use cursive::Printer;